refactor(step4): 코드리뷰 반영
- car 생성 검증 Car 도메인으로 이동 (기존 InputView)
- Cars.race 반환타입 RacingPhaseRecords 수정
