+++ src/main/kotlin/racing/domain/Car.kt
@@ -6,6 +6,12 @@ class Car(
     private val forwardCondition: ForwardCondition
 ) {
 
+    init {
+        if (name.length > 5) {
+            throw IllegalArgumentException(CAR_NAME_ERROR_MESSAGE)
+        }
+    }
+
     fun move(): RacingPhaseRecord {
         val forwardCondition = forwardCondition.getForwardCondition()
         if (forwardCondition) {
@@ -17,5 +23,6 @@ class Car(
 
     companion object {
         const val INITIAL_DISTANCE = 1
+        private const val CAR_NAME_ERROR_MESSAGE = "자동차 이름은 5자를 초과할 수 없습니다."
     }
 }
