+++ src/main/kotlin/racing/domain/Racing.kt
@@ -8,6 +8,5 @@ data class Racing(
     fun start(): List<RacingPhaseRecords> {
         return (0 until attemptCount)
             .map { cars.race() }
-            .map { RacingPhaseRecords(it) }
     }
 }
